requirement.txt
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
chromadbpip 
